test whether postgres user exists
So far, `pgsu` only tried the "sudo route" when it detected that it was
running on Ubuntu Linux.
It was reported that this test can fail when running on Windows
Subsystem for Linux.
Furthermore, this test might be too strict and exclude other Linux
distributions with similar postgresql setup (unverified).

Here, we switch from the check of the distribution name to a check
whether the `postgres` system user exists.

Furthermore, we introdyce the `try_sudo` and `sudo_user` constructor
arguments, which can be used to force `pgsu` to try the "sudo route" for
a given system user.