fix: change directory for sudo route
When `pgsu` tries to become the `postgres` user and run a `psql`
command, psql prints a warning [1] that it is not able to change directory:

```
WARNING  pgsu:__init__.py:368 could not change directory to "/path/to/c/w/d": Permission denied
```

Since the current implementation routes this to a python logger, this is
typically not seen by the user unless an exception is raised (or when
running tests).
It would be good to fix nevertheless.

[1] https://stackoverflow.com/questions/38470952/postgres-can-not-change-directory-in-ubuntu-14-04